Hidden files are not listed anymore When listing a directory, any hidden files under it will not be listed. To achieve this, the `iterate_dir` function was hooked. This function receives a directory context which has a callback function that is run for each entry under the listed directory. The hook replaces this callback with our own. Our fake callback simply calls the original callback (which we saved) only if the entry should not be hidden.
Additionally, the `path_lookupat` hook was replaced with a hook for the
`walk_component` function.
The hook's logic is identical, but unlike `path_lookupat`,
this function also catches all the intermediate steps of the lookup.
This was done because previously, we could access files under
a hidden directory because the files themselves were not hidden.

Tests were added for the new functionality.
Also, some timeouts were increased in the give_root tests because
on some runs they weren't enough.