Ghoul now hides itself from ftrace (#13) This is done by hooking the `print_trace_line` function. If the information for the trace line contains a function that belongs to us, we return a fake success indication. Fow now only the `function` and `function_graph` tracers are supported. I haven't seen any other tracer expose us, but if one turns out to, I will hide from it as well.
Hooking `print_trace_line` with ftrace turns out to be impossible.
It does not have ftrace's function prologue,
even though it's not marked as `notrace` in the source code.
Hooking it with kprobes failed as well,
because it caused an infinite recursion with kprobe's internal functions
needing to be traced as well.
The final solution I used was inline hooking (or splicing).
I borrowed the implementation from the suterusu rootkit:
https://github.com/mncoppola/suterusu

Currently, this new functionality does not have a test.
Testing this automatically seems complex
so for now it will need to be tested manually.

Two internal headers from the kernel were needed for this
so I included them in the repository.
Later I will need to automatically do this during the build process.

Also, some headers and definitions were reorganized.